@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <ImplicitUsings>enable</ImplicitUsings>
-    <TargetFrameworks>net7.0;net8.0</TargetFrameworks>
+	  <TargetFramework>net8.0</TargetFramework>
 	<IsPackable>true</IsPackable>
   </PropertyGroup>
 
